deep_research_app/backend/requirements.txt
# Deep Research Backend Requirements

# Microsoft Agent Framework (using beta version as stable not yet released)
agent-framework

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
python-dotenv>=1.0.0

# Azure dependencies (required by framework)
azure-identity>=1.15.0
azure-cosmos>=4.5.1
azure-mgmt-cognitiveservices>=13.5.0  # For Azure OpenAI deployment management

# Document processing dependencies
azure-ai-documentintelligence  # Document Intelligence SDK for PDF/DOCX
azure-storage-blob  # Blob storage for file uploads
python-docx>=1.1.0  # DOCX file processing
aiohttp>=3.9.0  # Async HTTP client for file downloads

# YAML support (required by framework for workflow definitions)
pyyaml>=6.0.1

# AI/ML dependencies
openai>=1.12.0
tavily-python>=0.3.0

# Export functionality
reportlab>=4.0.0
markdown>=3.5.0

# Additional utilities
aiofiles>=23.2.1
